Consolidate debug URLs into core urls.py. Allow schema-js to be accessed through setup middleware.
